add supermd
This is actually work that happenend in a branch but
because of an attempt to remove a bunch of vendored
libs from the history of the repo, merging doesn't work
anymore so I'm just merging everything as a single commit anyway.

See the `supermd` branch for the history of this work if you really
care, otherwise I would have squashed everything anyway.

Jokes on you git, your shitty interface in the end has not made me
do anything that I didn't want to do already.