Print an Error message when Wuffs returns an error
Wuffs is extremely strict with its parsing, so much so that it'll
reject images that other tools accept or emit over minor issues.
Some images I have get rejected even with no errors reported from tools
like exiftool, until I strip out all the Exif data.
In Zine, this manifests in the width and height attributes of your images
not being emitted, silently.

I've fixed this by printing out a warning with a suggestion of what action
the user can take. This is not the most ambitious of fixes, but it should
help nudge people in the right direction should they run into this issue.