add `test` step to zig build to run snapshot tests
under `tests/` you can find a list of websites (only one for now) that
will be built when running `zig build test`. the test will run git diff
on the entirety of `tests/` and expect nothing to have changed, failing
the build if that's not the case.

when the diff is not empty, the programmer will need to commit the new
changes in order to confirm that the change in behavior is wanted (ie
by committing the changes you update the golden snapshot).

Running tests implies `-Ddebug`. You can optionally passs `-Dscope` to
enable logging.

Adding new test websites doesn't require changing the main build script
as they will be automatically picked up. The only requirement for those
websites is to have `.debug = true` set, otherwise the `-Ddebug` flag
will not be available and the build will fail.

Lastly, to ensure that CI doesn't break, make sure to add an empty .keep
file in directories that have no content, but that must exist (eg `assets/`).