auto image size: add config flag and make errors hard errors
After the last release of Zine, a feature to automatically add
size attributes to local images was added to "unstable" Zine. The
reason for this practice is that it allows the browser to know the
size of all images immediately on page load, ensuring that no page
reflows are necessary (since the image will fit precisely in its
'predicted' box once it gets downloaded).

But this feature has two downsides:
- Browsers have confusing behavior when `width` and `height` attributes
  are set on images, causing the images to be stretched unless a css
  "patch" is applied (`img { height: auto; }`).
- The library that we use to parse local images is Wuffs and it
  sometimes refuses to load normal images (e.g. photos taken with a
  phone or even just screenshots), probably because of malformed
  metadata.

While the value of having a website that doesn't need to be reflowed on
load is undoubtably real, these two issues combined mean that a user
must be interested in the active pursuit of a reflow-less website,
especially because they will need occasionally to fix their images in
order for Wuffs to accept them.

For this reason the feature is now being put behind a flag.

Add `image_size_attributes = true` to your `build.zig` to enable it.
Once the feature is enabled, image parsing errors from Wuffs will become
hard errors in Zine.

I'm open to reconsider having this feature enabled by default if a
strong solution to the second flaw is found. In the meantime I'll
add a good section on the official website to explain to people why
it might be worth enabling the flag.