Ignore footnotes ids when handlings content sections
Footnotes produce a reference that can be linked to and will show up in
the `ids` part of a page's Ast (see #92).

Footnotes have no directive attached to them, as their referenced
sections are added implicitly when they are processed by the parser.

The `contentSection` related methods did assume that all nodes in `ids`
are proper sections, so having footnotes would break layouts that use
such an API, e.g. `$page.contentSections()`.

This commit updates those implementations to expect nodes to not
necessarily have a directive.

Additionally, the logic of `hasContentSection` is aligned with the
behavior of `contentSection(id)`, so that returning a true value will
not trigger an error when getting the content section.
This would have been possible before id the node with that name is not a
`section`, however it might have been that, in practice, this was never
the case.

I added a footnote to the sections snapshot to show that rendering this
page does no longer break.