significantly improved linking
linking to things (pages, assets, ...) is complicated to implement
correctly, especially when dealing with multilingual websites and all
the related customizations (custom urls, fetching pages from other
variants through scripty, etc).

this commit moves link composition in supermd from the analysis phase
to the rendering phase in order to account for situations where a page
is rendered in a context other than its natural one. an example of that
is rendering the contents of a page inside an RSS feed.

specifically to help with RSS feeds, Zine will automatically use full
links when rendering a page fetched via scripty.

the code is not yet fully correct as it's missing some normalization
in certain places and some usage of std.fs.path.fmtJoin will produce
wrong results on windows. those are left as todos for a subsequent
commit.

with this commit the official zig website builds and the diff only
contains positive changes.