fix string table misusage
turns out that trying to be clever with string tables requires being
clevererer. when trying to create new entries based on existing ones
one needs to be careful in not holding a pointer across a resize, which
sometimes is hard because that's the only way you can express the
self-referentialness. the solution is simply to call ensure unused
capacity at the right time