implement memory / disk modes
This feature is motivated by the existence of the live server.

In previous iterations of Zine, and in most other site generators,
the live server component is bolted on top of the normal build pipeline,
meaning that, on change, the site is rebuilt and installed in a location
on disk, which is what the live server offers over http.

It's actually pretty obvious that the optimal thing would be to not
install any file anywhere at all, and instead serve all generated files
from memory. Not only is this more efficient, but it also solves easily
one annoying problem that tends to plague from-disk live servers: stale
files.

In previous iterations of Zine (and some other SSGs) stale files would
not be deleted from the output dir, resulting in confusing behavior when
the user refreshed a page that was moved, seeing stale content with no
clear indication that this is what's happening.

Additionally, leveraging the new Zine architecture we are able to tell
users when an asset file exists on disk, but it would be a 404 because
it was never referenced anywhere (and thus would not be installed in the
output dir).

Now the main site building process accepts a parameter that defines if
the build is meant to be installed to a disk location or kept in memory.

This change might have introduced regressions in path prefix handling
and it's still missing one big feature for the live server: keeping
errors in memory so that they can be shown to clients that load after
the build has already been computed. This will be implemented in a
subsequent commit.