reintroduce linux and windows file watchers
the interface has been modified slightly, mainly
dropping the distinction between input and output
directories (now we don't even have output files
in the live server!) and removing the whole idea
of a 'reloader'. The watchers will instead call
a known function in Debouncer which in turn will
trigger an event in the main serve thread.