reintroduce error reporting in the live server
This is an initial rough implementation of error reporting in the
browser.

Two major innovations compared to the previous iteration:

- The build state is persisted, meaning that we can show errors to
  sockets that connect after the fact. This makes the whole system
  infinitely more robust than in the past because you will always
  see the error even if you refresh the page, while previously, if
  you missed the initial notification, you would see nothing till the
  next rerun.
- We don't spam terminal escape codes anymore. Now errors are rendered
  separately for the terminal vs for any other subsequent reporting (
  which for now only means the live server). The implementation is not
  great at the moment, you can see a ton of duplicated lines and it
  clearly would benefit from a design pass first (i.e. trying to answer
  the question of what would be the optimal way of displaying errors in
  the browser), and then a better implementation pass second :^)

The design is not super polished but it works well for now:

If any error is present, which means that pages were not rendered (and
maybe not even analyzed), then trying to get a page will return an error
HTML page that will setup the websocket connection and show dynamically
the last build errors that happened.

When multiple errors are present, those will be sent as multiple messages
which the client concatenates into a single view. To avoid spam from
template rendering errors, only one of those gets shown at a time.

Lastly, this setup leverages the fact that the client is told to refresh
the page unconditionally at every rebuild (since we don't have caching).
Once we implement proper caching this system will have to be tweaked to
also notify currently connected sockets that are not going to reload.