reintrouduce subdirectory support in layouts
when reimplementing zine I temporarily regressed the ability to put
layouts and templates inside of subdirs. this restriction has now
been lifted.

This implementation allows you to create subtrees inside of both
`layouts/` and `templates/`. When a template defines a parent
template that they extend, the template will be searched inside of
`templates/`, meaning that all template paths are relative to
'layouts/templates/'.

this commit also contains some improvements to the superhtml docgen
tool.