improve katek integration
previously we were using a script provided directly by the katek project
to integrate with our rendered equations, but that script did not handle
display mode out of the box. further investigation showed that the script
expected us to do weird shit with the script type attribute and so we
replace it with another implementation that is saner overall.

link to the official implementation:
https://github.com/KaTeX/KaTeX/tree/main/contrib/mathtex-script-type

note that this integration is temporary as the ideal way of implementing
latex equations would be to do the parsing work at build time, and to
output mathml so that clients don't need to have JS enabled, which is
why I'm not bothering with a noscript fallback for now.