Add IPv6 support to the development server
This adds support for IPv6.  The server already uses the APIs needed for
this, so this is just a matter of handling the string parsing properly.

Since the `--host` argument supports specifying the port by appending
it to the address with a colon (e.g. `:8080`), I opted to require
surrounding IPv6 addresses in square brackets (e.g. `[::1]`) so as not
to break backwards compatibility.  There is precedent to square bracket
notation of IPv6 addresses in e.g. URLs.